Fix isDelegated method call

+++ contracts/interfaces/IChefData.sol
@@ -5,7 +5,7 @@ pragma solidity 0.6.12;
 interface IOptionChefData {
     function uIds(uint k) external view returns (uint);
     function ids(uint k) external view returns (uint);
-    function optionType(uint k) external view returns (uint);
+    function optionType(uint k) external view returns (uint8);
 
     function setuid(uint k, uint v) external;
     function setid(uint k, uint v) external;
